Add metadata to static pages
